requirements.txt
Flask==2.3.3
flask
flask-mysql-connector
